requirements.txt
streamlit
pillow
numpy
gdown
tensorflow
